docs: update examples
